move ktisis classes from `Structs` to `Posing`
