+++ Ktisis/Configuration.cs
@@ -9,10 +9,10 @@ using Dalamud;
 using Dalamud.Configuration;
 using Dalamud.Game.ClientState.Keys;
 
+using Ktisis.Posing;
 using Ktisis.Services;
 using Ktisis.Interface;
 using Ktisis.Structs.Bones;
-using Ktisis.Structs.Poses;
 using Ktisis.Structs.Actor.Equip.SetSources;
 using static Ktisis.Data.Files.AnamCharaFile;
