+++ Ktisis/Data/Serialization/Converters/NumericConverters.cs
@@ -4,7 +4,7 @@ using System.Text.Json;
 using System.Text.Json.Serialization;
 using System.Globalization;
 
-using Ktisis.Structs.Poses;
+using Ktisis.Posing;
 
 namespace Ktisis.Data.Serialization.Converters {
