+++ Ktisis/Interface/Overlay/Skeleton.cs
@@ -5,9 +5,9 @@ using ImGuiNET;
 
 using ImGuizmoNET;
 
+using Ktisis.Posing;
 using Ktisis.Services;
 using Ktisis.Structs.Actor;
-using Ktisis.Structs.Bones;
 using Ktisis.Library.Extensions;
 
 namespace Ktisis.Interface.Overlay {
