+++ Ktisis/Interop/Hooks/PoseHooks.cs
@@ -7,9 +7,9 @@ using Dalamud.Game.ClientState.Objects.Types;
 using FFXIVClientStructs.Havok;
 using FFXIVClientStructs.FFXIV.Client.Graphics.Render;
 
+using Ktisis.Posing;
 using Ktisis.Services;
 using Ktisis.Structs.Actor;
-using Ktisis.Structs.Poses;
 using Ktisis.Library.Extensions;
 
 namespace Ktisis.Interop.Hooks {
